Refactored base pathfinding classes
