--- Assets/CodeBase/Pathfinding/AI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b3712b4184fb54444974d7e56c957a21
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
