+++ Assets/CodeBase/Pathfinding/Base Pathfinding.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d26dd2af1ca0f004eae37e59e409f692
+guid: 6031282c11323f7449f4a6d09fb95cf7
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
