+++ Assets/CodeBase/Pathfinding/Base Pathfinding/BaseNode.cs
@@ -1,12 +1,16 @@
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
-public abstract class BaseNode<T>
+namespace Pathfinding.BasePathfinding
 {
-    public T Value { get; private set; }
-    public BaseNode(T value)
+    public abstract class BaseNode<T>
     {
-        Value = value;
-    }
+        public T Value { get; private set; }
 
-    public abstract List<BaseNode<T>> GetNeighbours();
+        public BaseNode(T value)
+        {
+            Value = value;
+        }
+
+        public abstract List<BaseNode<T>> GetNeighbours();
+    }
 }
