+++ Assets/CodeBase/Pathfinding/Base Pathfinding/BasePathfinder.cs
@@ -1,22 +1,28 @@
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
 namespace Pathfinding.BasePathfinding
 {
     public abstract class BasePathfinder<T>
     {
-        public PathfinderStatus Status { get; private set; } = PathfinderStatus.NOT_INITIALIZED;
-        public BaseNode<T> StartNode { get; private set; }
-        public BaseNode<T> TargetNode { get; private set; }
-        public PathfinderNode<T> CurrentNode { get; private set; }
-
         public delegate float CostFunction(T a, T b);
         public CostFunction HeuristicCost { get; set; }
         public CostFunction NodeTraversalCost { get; set; }
 
-        protected List<PathfinderNode<T>> OpenList = new List<PathfinderNode<T>>();
-        protected List<PathfinderNode<T>> CloseList = new List<PathfinderNode<T>>();
+        public PathfinderStatus Status { get; private set; } = PathfinderStatus.NOT_INITIALIZED;
 
-        public bool Initialize(BaseNode<T> startNode, BaseNode<T> targetNode)
+        public BaseNode<T> Start { get; private set; }
+        public BaseNode<T> Goal { get; private set; }
+
+        public PathfinderNode<T> CurrentNode { get; private set; }
+
+        public delegate void DelegateNoArgument();
+        public DelegateNoArgument onFailure;
+        public DelegateNoArgument onSuccess;
+
+        protected List<PathfinderNode<T>> OpenList = new List<PathfinderNode<T>>();
+        protected List<PathfinderNode<T>> ClosedList = new List<PathfinderNode<T>>();
+
+        public bool Initialize(BaseNode<T> start, BaseNode<T> goal)
         {
             if (Status == PathfinderStatus.RUNNING)
             {
@@ -24,31 +30,31 @@ namespace Pathfinding.BasePathfinding
             }
 
             Reset();
-            StartNode = startNode;
-            TargetNode = targetNode;
+
+            Start = start;
+            Goal = goal;
+
+            float H = HeuristicCost(Start.Value, Goal.Value);
+
+            PathfinderNode<T> root = new PathfinderNode<T>(Start, null, 0f, H);
+
+            OpenList.Add(root);
+
+            CurrentNode = root;
 
             Status = PathfinderStatus.RUNNING;
+
             return true;
         }
-
-        public void Reset()
-        {
-            if (Status == PathfinderStatus.RUNNING)
-            {
-                return;
-            }
-
-            CurrentNode = null;
-            Status = PathfinderStatus.NOT_INITIALIZED;
-        }
-
+        
         public PathfinderStatus Step()
         {
-            CloseList.Add(CurrentNode);
+            ClosedList.Add(CurrentNode);
 
             if (OpenList.Count == 0)
             {
                 Status = PathfinderStatus.FAILURE;
+                onFailure?.Invoke();
                 return Status;
             }
 
@@ -56,49 +62,67 @@ namespace Pathfinding.BasePathfinding
 
             OpenList.Remove(CurrentNode);
 
-            if (EqualityComparer<T>.Default.Equals(CurrentNode.ChildNode.Value, TargetNode.Value))
+            if (EqualityComparer<T>.Default.Equals(
+                CurrentNode.Location.Value, Goal.Value))
             {
-                Status = PathfinderStatus.SUCCSESS;
+                Status = PathfinderStatus.SUCCESS;
+                onSuccess?.Invoke();
                 return Status;
             }
 
-            List<BaseNode<T>> neighbours = CurrentNode.ChildNode.GetNeighbours();
+            List<BaseNode<T>> neighbours = CurrentNode.Location.GetNeighbours();
 
-            foreach(BaseNode<T> node in neighbours)
+            foreach (BaseNode<T> cell in neighbours)
             {
-                AlgorithmImplementation(node);
+                AlgorithmSpecificImplementation(cell);
             }
 
             Status = PathfinderStatus.RUNNING;
             return Status;
         }
 
-        private PathfinderNode<T> GetLeastCostNode(List<PathfinderNode<T>> list)
+        protected void Reset()
+        {
+            if (Status == PathfinderStatus.RUNNING)
+            {
+                return;
+            }
+
+            CurrentNode = null;
+
+            OpenList.Clear();
+            ClosedList.Clear();
+
+            Status = PathfinderStatus.NOT_INITIALIZED;
+        }
+
+        protected PathfinderNode<T> GetLeastCostNode(List<PathfinderNode<T>> myList)
         {
             int best_index = 0;
-            float best_priority = list[0].TotalCost;
-            for (int i = 1; i < list.Count; i++)
+            float best_priority = myList[0].Fcost;
+            for (int i = 1; i < myList.Count; i++)
             {
-                if (best_priority > list[i].TotalCost)
+                if (best_priority > myList[i].Fcost)
                 {
-                    best_priority = list[i].TotalCost;
+                    best_priority = myList[i].Fcost;
                     best_index = i;
                 }
             }
-            PathfinderNode<T> n = list[best_index];
+
+            PathfinderNode<T> n = myList[best_index];
             return n;
         }
 
-        protected int GetNodePositionInList(List<PathfinderNode<T>> list, T targetNode)
+        protected int IsInList(List<PathfinderNode<T>> myList, T cell)
         {
-            for(int i=0; i<list.Count; i++)
+            for (int i = 0; i < myList.Count; ++i)
             {
-                if (EqualityComparer<T>.Default.Equals(list[i].ChildNode.Value, targetNode))
+                if (EqualityComparer<T>.Default.Equals(myList[i].Location.Value, cell))
                     return i;
             }
             return -1;
         }
 
-        protected abstract void AlgorithmImplementation(BaseNode<T> node);
+        abstract protected void AlgorithmSpecificImplementation(BaseNode<T> cell);
     }
 }
