+++ Assets/CodeBase/Pathfinding/Base Pathfinding/PathfinderStatus.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4cfd1fa0983b87545a990d3eaf06eecc
+guid: 1f9f596e6bf4a244f95a0269b48107d4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
