+++ Assets/CodeBase/Pathfinding/Pathfinding Algorithms/AStarPathfinder.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e69ada9276038d94ba84bcc04f330916
+guid: 42092033e8cf0ed4cae52a14c8d6266e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
