+++ Assets/CodeBase/Pathfinding/Pathfinding Algorithms/DijkstraPathfinder.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 52ffc36c948480143bd4b58f894501f9
+guid: 156dd43f11db01b43ba98aec62b3ec73
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
